requirements.txt
langchain
langchain-openai
langchain-community
azure-search-documents
azure-monitor-opentelemetry
azure-identity
azure-storage-blob
azure-ai-evaluation
openai
streamlit
python-dotenv
pytest
ragas
scikit-learn
Pillow
httpx
pandas
mlflow<=3.5.0
azureml-mlflow
azure-ai-contentsafety
pytest
pytest-asyncio
kaggle
chromadb
python-json-logger
